# openapi.yaml
openapi: 3.1.0
info:
  title: 포토스팟 추천 API
  description: 퍼스널컬러·체형 기반 사진 명소 추천. 소셜 로그인(/v1/auth/*)으로 받은 액세스 토큰을 Bearer로 보낸다.
  version: 1.2.0
paths:
...
  /v1/social/instagram/callback:
    get:
      tags:
      - growth
      summary: Instagram Callback
      operationId: instagram_callback_v1_social_instagram_callback_get
      parameters:
      - name: state
        in: query
        required: true
        schema:
          type: string
          maxLength: 256
          title: State
      - name: code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 4096
          - type: 'null'
          title: Code
      - name: error
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Error
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/feedback:
    post:
      tags:
      - growth
      summary: Submit Feedback
      operationId: submit_feedback_v1_feedback_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackSubmission'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/me/feedback/{feedback_id}:
    delete:
      tags:
      - growth
      summary: Delete Feedback
      operationId: delete_feedback_v1_me_feedback__feedback_id__delete
      parameters:
      - name: feedback_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Feedback Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/admin/feedback/{feedback_id}:
    patch:
      tags:
      - growth
      summary: Review Feedback
      operationId: review_feedback_v1_admin_feedback__feedback_id__patch
      parameters:
      - name: feedback_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Feedback Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackReview'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/admin/feedback/triage:
    post:
      tags:
      - growth
      summary: Analyze Feedback
      operationId: analyze_feedback_v1_admin_feedback_triage_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriageBatch'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/admin/triages/{triage_id}:
    patch:
      tags:
      - growth
      summary: Decide Triage
      operationId: decide_triage_v1_admin_triages__triage_id__patch
      parameters:
      - name: triage_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Triage Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriageDecision'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/admin/campaigns:
    get:
      tags:
      - growth
      summary: Campaigns
      operationId: campaigns_v1_admin_campaigns_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      tags:
      - growth
      summary: Create Campaign
      operationId: create_campaign_v1_admin_campaigns_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/admin/campaigns/{campaign_id}:
    patch:
      tags:
      - growth
      summary: Approve Campaign
      operationId: approve_campaign_v1_admin_campaigns__campaign_id__patch
      parameters:
      - name: campaign_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Campaign Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignApproval'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sponsorships:
    get:
      tags:
      - growth
      summary: Sponsorships
      operationId: sponsorships_v1_sponsorships_get
      parameters:
      - name: lat
        in: query
        required: true
        schema:
          type: number
          maximum: 39
          minimum: 33
          title: Lat
      - name: lng
        in: query
        required: true
        schema:
          type: number
          maximum: 132
          minimum: 124
          title: Lng
      - name: radius_m
        in: query
        required: false
        schema:
          type: integer
          maximum: 50000
          minimum: 500
          default: 5000
          title: Radius M
      - name: place_group
        in: query
        required: false
        schema:
          type: string
          pattern: ^(all|cafe|travel|festival)$
          default: all
          title: Place Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sponsorships/{campaign_id}/events:
    post:
      tags:
      - growth
      summary: Record Event
      operationId: record_event_v1_sponsorships__campaign_id__events_post
      parameters:
      - name: campaign_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Campaign Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignEvent'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CampaignApproval:
      properties:
        approved:
          type: boolean
          title: Approved
      type: object
      required:
      - approved
      title: CampaignApproval
    CampaignEvent:
      properties:
        kind:
          type: string
          enum:
          - impression
          - click
          title: Kind
      type: object
      required:
      - kind
      title: CampaignEvent
    CampaignIn:
      properties:
        place_id:
          type: string
          format: uuid
          title: Place Id
        sponsor:
          type: string
          maxLength: 80
          minLength: 1
          title: Sponsor
        headline:
          type: string
          maxLength: 100
          minLength: 1
          title: Headline
        placement:
          type: string
          enum:
          - banner
          - priority
          title: Placement
        starts_at:
          type: string
          format: date-time
          title: Starts At
        ends_at:
          type: string
          format: date-time
          title: Ends At
      type: object
      required:
      - place_id
      - sponsor
      - headline
      - placement
      - starts_at
      - ends_at
      title: CampaignIn
...
    FeedbackOut:
      properties:
        category:
          type: string
          enum:
          - bug
          - idea
          - place
          - other
          title: Category
        message:
          type: string
          maxLength: 2000
          minLength: 5
          title: Message
        ai_consent:
          type: boolean
          title: Ai Consent
          default: false
        id:
          type: string
          format: uuid
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        reviewed_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Reviewed Text
      type: object
      required:
      - category
      - message
      - id
      - created_at
      title: FeedbackOut
    FeedbackReview:
      properties:
        reviewed_text:
          type: string
          maxLength: 2000
          minLength: 5
          title: Reviewed Text
      type: object
      required:
      - reviewed_text
      title: FeedbackReview
    FeedbackSubmission:
      properties:
        category:
          type: string
          enum:
          - bug
          - idea
          - place
          - other
          title: Category
        message:
          type: string
          maxLength: 2000
          minLength: 5
          title: Message
        ai_consent:
          type: boolean
          title: Ai Consent
          default: false
      type: object
      required:
      - category
      - message
      title: FeedbackSubmission
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TriageBatch:
      properties:
        feedback_ids:
          items:
            type: string
            format: uuid
          type: array
          maxItems: 20
          minItems: 1
          title: Feedback Ids
      type: object
      required:
      - feedback_ids
      title: TriageBatch
    TriageDecision:
      properties:
        status:
          type: string
          enum:
          - approved
          - planned
          - done
          - dismissed
          title: Status
        note:
          type: string
          maxLength: 1000
          minLength: 3
          title: Note
      type: object
      required:
      - status
      - note
      title: TriageDecision
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError